+++ docs/index.md
@@ -13,8 +13,10 @@ Questa cartella raccoglie una serie di note operative e di design pensate per ac
 - [Training del modello gerarchico](training_hier_model.md)
 - [Pre-training TAPT/MLM](training_tapt.md)
 - [Comando `convert`](cli_convert.md)
-- [Comando `train`](cli_train.md)
-- [Comando `validate`](cli_validate.md)
+- [Comando `extract`](cli_extract.md)
+- [Comando `evaluate`](cli_evaluate.md)
+- [Comando `pack`](cli_pack.md)
+- [Script di training](cli_train.md)
 - [Pipeline di inferenza end-to-end](inference_pipeline.md)
 - [Servizio FastAPI](service_api.md)
 - [Gestione del knowledge pack](core_knowledge_pack.md)
