+++ src/robimb/props/cli.py
@@ -1,27 +1,34 @@
 import argparse
 from pathlib import Path
-from .unpack import convert_monolith_to_folders
-from ..registry import pack_folders_to_monolith
 
-def main():
-    p = argparse.ArgumentParser(prog="robimb", description="Registry/Extractors converter")
-    sub = p.add_subparsers(dest="cmd", required=True)
+from ..cli.extract import extract_command
+from ..cli.pack import pack_command
 
-    c = sub.add_parser("convert", help="Monolith (registry+extractors) -> properties folders")
-    c.add_argument("--in-registry", required=True, type=Path)
-    c.add_argument("--in-extractors", required=True, type=Path)
-    c.add_argument("--out-dir", required=True, type=Path)
 
-    k = sub.add_parser("pack", help="properties folders -> monolith")
-    k.add_argument("--in-dir", required=True, type=Path)
-    k.add_argument("--out-registry", required=True, type=Path)
-    k.add_argument("--out-extractors", required=True, type=Path)
+def main() -> None:
+    parser = argparse.ArgumentParser(prog="robimb", description="Registry/Extractors converter")
+    sub = parser.add_subparsers(dest="cmd", required=True)
 
-    a = p.parse_args()
-    if a.cmd == "convert":
-        convert_monolith_to_folders(a.in_registry, a.in_extractors, a.out_dir)
-        print(f"OK: properties written to {a.out_dir}")
+    cmd_extract = sub.add_parser("convert", help="Monolith (registry+extractors) -> properties folders")
+    cmd_extract.add_argument("--in-registry", required=True, type=Path)
+    cmd_extract.add_argument("--in-extractors", required=True, type=Path)
+    cmd_extract.add_argument("--out-dir", required=True, type=Path)
+
+    cmd_pack = sub.add_parser("pack", help="properties folders -> monolith")
+    cmd_pack.add_argument("--in-dir", required=True, type=Path)
+    cmd_pack.add_argument("--out-registry", required=True, type=Path)
+    cmd_pack.add_argument("--out-extractors", required=True, type=Path)
+
+    args = parser.parse_args()
+    if args.cmd == "convert":
+        extract_command(
+            in_registry=args.in_registry,
+            in_extractors=args.in_extractors,
+            out_dir=args.out_dir,
+        )
     else:
-        pack_folders_to_monolith(a.in_dir, a.out_registry, a.out_extractors)
-        print(f"OK: registry -> {a.out_registry}")
-        print(f"OK: extractors -> {a.out_extractors}")
+        pack_command(
+            properties_root=args.in_dir,
+            out_registry=args.out_registry,
+            out_extractors=args.out_extractors,
+        )
